move all tests into test folder
