+++ src/components/gcds-grid/test/gcds-grid.spec.ts
@@ -1,5 +1,5 @@
 import { newSpecPage } from '@stencil/core/testing';
-import { GcdsGrid} from './gcds-grid';
+import { GcdsGrid} from '../gcds-grid';
 
 describe('gcds-grid', () => {
   it('renders', async () => {
